ZENKO-3368: reorganize auth backend files
