+++ lib/auth/auth.js
@@ -10,10 +10,10 @@ const constants = require('../constants');
 const constructStringToSignV2 = require('./v2/constructStringToSign');
 const constructStringToSignV4 = require('./v4/constructStringToSign');
 const convertUTCtoISO8601 = require('./v4/timeUtils').convertUTCtoISO8601;
-const vaultUtilities = require('./in_memory/vaultUtilities');
-const backend = require('./in_memory/Backend');
-const validateAuthConfig = require('./in_memory/validateAuthConfig');
-const AuthLoader = require('./in_memory/AuthLoader');
+const vaultUtilities = require('./backends/in_memory/vaultUtilities');
+const inMemoryBackend = require('./backends/in_memory/Backend');
+const validateAuthConfig = require('./backends/in_memory/validateAuthConfig');
+const AuthLoader = require('./backends/in_memory/AuthLoader');
 const Vault = require('./Vault');
 
 let vault = null;
@@ -215,7 +215,7 @@ module.exports = {
         constructStringToSignV2,
     },
     inMemory: {
-        backend,
+        backend: inMemoryBackend,
         validateAuthConfig,
         AuthLoader,
     },
