+++ lib/auth/backends/in_memory/AuthLoader.js
@@ -3,7 +3,7 @@ const glob = require('simple-glob');
 const joi = require('@hapi/joi');
 const werelogs = require('werelogs');
 
-const ARN = require('../../models/ARN');
+const ARN = require('../../../models/ARN');
 
 /**
  * Load authentication information from files or pre-loaded account
