+++ lib/auth/backends/in_memory/Indexer.js
