+++ lib/auth/backends/in_memory/validateAuthConfig.js
