+++ lib/auth/backends/in_memory/vaultUtilities.js
